# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Image:
    description: An image is what is processed in the application.
    properties:
      id:
        $ref: '#/definitions/UUID'
      resulted_location:
        description: the resulted location for this image
        type: string
        x-go-name: ResultedLocation
      resulted_name:
        description: the resulted name for this image
        type: string
        x-go-name: ResultedName
      uploaded_location:
        description: the uploaded location for this image
        type: string
        x-go-name: UploadedLocation
      uploaded_name:
        description: the uploaded name for this image
        type: string
        x-go-name: UploadedName
    required:
    - uploaded_name
    - uploaded_location
    title: Image contains details of a single request.
    type: object
    x-go-package: github.com/alisavch/image-service/internal/models
  UUID:
    description: |-
      A UUID is a 128 bit (16 byte) Universal Unique IDentifier as defined in RFC
      4122.
    items:
      format: uint8
      type: integer
    type: array
    x-go-package: github.com/google/uuid
  User:
    description: A user is the security principal for this application.
    properties:
      id:
        $ref: '#/definitions/UUID'
      password:
        description: the password for this user
        type: string
        x-go-name: Password
      username:
        description: the username for this user
        type: string
        x-go-name: Username
    required:
    - username
    - password
    title: User contains information about user.
    type: object
    x-go-package: github.com/alisavch/image-service/internal/models
host: localhost:8080
info:
  description: Documentation of our awesome API
  title: Golang SwaggerUI image-service
  version: 1.0.0
paths:
  /api/compress:
    post:
      description: Receives an image from an input form and compresses it, also allows
        you to enter width in the query string.
      operationId: compress
      parameters:
      - in: query
        name: width
        type: integer
      - in: body
        name: uploadFile
        required: true
        schema:
          $ref: '#/definitions/Image'
      responses:
        "200":
          description: successful operation
        "401":
          description: unauthorized user
        "500":
          description: internal server error
      summary: Compresses the image.
      tags:
      - compress
  /api/compress/{compressedID}:
    get:
      description: Downloads the compressed image and original if required.
      operationId: findCompressedImage
      parameters:
      - description: compressedID to filter by id
        in: path
        name: compressedID
        required: true
        type: integer
      - description: If true - the original will be saved
        in: query
        name: original
        type: boolean
      responses:
        "200":
          description: successful operation
        "401":
          description: unauthorized user
        "500":
          description: internal server error
      summary: Finds the compressed image.
      tags:
      - findCompressedImage
  /api/convert:
    post:
      description: Receives an image from an input form and converts it PNG to JPG
        and vice versa.
      operationId: convert
      parameters:
      - in: body
        name: uploadFile
        required: true
        schema:
          $ref: '#/definitions/Image'
      responses:
        "200":
          description: successful operation
        "401":
          description: unauthorized user
        "500":
          description: internal server error
      summary: Converts the image.
      tags:
      - convert
  /api/convert/{convertedID}:
    get:
      description: Downloads the converted image and original if required.
      operationId: findConvertedImage
      parameters:
      - description: convertedID to filter by id
        in: path
        name: convertedID
        required: true
        type: integer
      - description: If true - the original will be saved
        in: query
        name: original
        type: boolean
      responses:
        "200":
          description: successful operation
        "401":
          description: unauthorized user
        "500":
          description: internal server error
      summary: Finds the converted image.
      tags:
      - findConvertedImage
  /api/history:
    get:
      description: Lists all queries created by user.
      operationId: history
      responses:
        "200":
          description: successful operation
        "401":
          description: unauthorized user
        "500":
          description: internal server error
      summary: Finds users history.
      tags:
      - history
  /api/sign-in:
    post:
      description: Only authorized user has access.
      operationId: sign-in
      parameters:
      - description: the user to create
        in: body
        name: user
        schema:
          $ref: '#/definitions/User'
      responses:
        "200":
          description: successful operation
        "401":
          description: unauthorized user
      summary: Authorizes the user.
      tags:
      - sign-in
  /api/sign-up:
    post:
      description: Could be any user.
      operationId: sign-up
      parameters:
      - description: the user to create
        in: body
        name: user
        schema:
          $ref: '#/definitions/User'
      responses:
        "201":
          description: user created successfully
        "400":
          description: bad request
        "409":
          description: user already exists
        "500":
          description: internal server error
      summary: Registers a user.
      tags:
      - sign-up
produces:
- application/json
schemes:
- http
securityDefinitions:
  basic:
    type: basic
swagger: "2.0"